refactoring Cart state management
