+++ src/components/Cart.jsx
@@ -1,18 +1,26 @@
-import { useContext } from "react";
+import { useContext, useState, useEffect } from "react";
 import { Birds } from "../Context";
 import Bonus from "./Bonus";
 
-const Cart = ({ totalPrice, removeItem }) => {
-  const { birds } = useContext(Birds);
+const Cart = () => {
+  const { birds, setBirds } = useContext(Birds);
 
-  // console.log(useContext(Birds));
+  const [totalPrice, setTotalPrice] = useState();
+
+  function removeItem(id) {
+    setBirds(birds.filter((item) => item.id !== id));
+  }
+
+  useEffect(() => {
+    setTotalPrice(birds.reduce((accu, { amount }) => accu + amount, 0));
+  }, [birds]);
+
   return (
     <>
       <h2>Cart</h2>
       <h4>Discount: {birds.length < 3 ? 0 : 10}%</h4>
       <h5>
-        Total: ${" "}
-        {birds.length < 3 ? totalPrice : (totalPrice -= totalPrice * 0.1)}
+        Total: $ {birds.length < 3 ? totalPrice : totalPrice - totalPrice * 0.1}
       </h5>
       {birds.map(({ id, name, amount }) => {
         return (
